// src/ui/wave8ArtManifest.generated.ts
export const WAVE8_IMAGES = {
  chronicle_initial_a: {"url": "assets/wave8/ornaments/chronicle_initial_a.png", "width": 128, "height": 128},
  chronicle_initial_b: {"url": "assets/wave8/ornaments/chronicle_initial_b.png", "width": 128, "height": 128},
  chronicle_initial_c: {"url": "assets/wave8/ornaments/chronicle_initial_c.png", "width": 128, "height": 128},
  frame_chronicle_page: {"url": "assets/wave8/frames/frame_chronicle_page.png", "width": 512, "height": 640},
  frame_minimap: {"url": "assets/wave8/frames/frame_minimap.png", "width": 256, "height": 256},
  frame_objective_drawer: {"url": "assets/wave8/frames/frame_objective_drawer.png", "width": 256, "height": 512},
  frame_overlay_legend: {"url": "assets/wave8/frames/frame_overlay_legend.png", "width": 192, "height": 256},
  frame_petition: {"url": "assets/wave8/frames/frame_petition.png", "width": 384, "height": 256},
  frame_season_ledger: {"url": "assets/wave8/frames/frame_season_ledger.png", "width": 512, "height": 384},
  icon_alert_bell_bad: {"url": "assets/wave8/icons/icon_alert_bell_bad.png", "width": 96, "height": 96},
  icon_alert_bell_info: {"url": "assets/wave8/icons/icon_alert_bell_info.png", "width": 96, "height": 96},
  icon_alert_bell_threat: {"url": "assets/wave8/icons/icon_alert_bell_threat.png", "width": 96, "height": 96},
  icon_overlay_danger: {"url": "assets/wave8/icons/icon_overlay_danger.png", "width": 96, "height": 96},
  icon_overlay_finance: {"url": "assets/wave8/icons/icon_overlay_finance.png", "width": 96, "height": 96},
  icon_overlay_food: {"url": "assets/wave8/icons/icon_overlay_food.png", "width": 96, "height": 96},
  icon_overlay_labour: {"url": "assets/wave8/icons/icon_overlay_labour.png", "width": 96, "height": 96},
  icon_overlay_rights: {"url": "assets/wave8/icons/icon_overlay_rights.png", "width": 96, "height": 96},
  icon_overlay_transport: {"url": "assets/wave8/icons/icon_overlay_transport.png", "width": 96, "height": 96},
  icon_overlay_water: {"url": "assets/wave8/icons/icon_overlay_water.png", "width": 96, "height": 96},
  icon_overlay_zone: {"url": "assets/wave8/icons/icon_overlay_zone.png", "width": 96, "height": 96},
  keyart_mode_select: {"url": "assets/wave8/keyart/keyart_mode_select.png", "width": 1280, "height": 720},
  keyart_title_bg: {"url": "assets/wave8/keyart/keyart_title_bg.png", "width": 1920, "height": 1080},
  keyart_title_emblem: {"url": "assets/wave8/keyart/keyart_title_emblem.png", "width": 512, "height": 512},
  loading_1315_famine: {"url": "assets/wave8/keyart/loading_1315_famine.png", "width": 1920, "height": 1080},
  loading_1337_war: {"url": "assets/wave8/keyart/loading_1337_war.png", "width": 1920, "height": 1080},
  loading_1348_plague: {"url": "assets/wave8/keyart/loading_1348_plague.png", "width": 1920, "height": 1080},
  pause_badge: {"url": "assets/wave8/time/pause_badge.png", "width": 128, "height": 48},
  pause_vignette: {"url": "assets/wave8/time/pause_vignette.png", "width": 1024, "height": 640},
  seal_petition_accept: {"url": "assets/wave8/ornaments/seal_petition_accept.png", "width": 64, "height": 64},
  seal_petition_price: {"url": "assets/wave8/ornaments/seal_petition_price.png", "width": 64, "height": 64},
  seal_petition_reject: {"url": "assets/wave8/ornaments/seal_petition_reject.png", "width": 64, "height": 64},
  season_event_harvest: {"url": "assets/wave8/time/season_event_harvest.png", "width": 32, "height": 32},
  season_event_market_day: {"url": "assets/wave8/time/season_event_market_day.png", "width": 32, "height": 32},
  season_event_period_end: {"url": "assets/wave8/time/season_event_period_end.png", "width": 32, "height": 32},
  season_event_sow: {"url": "assets/wave8/time/season_event_sow.png", "width": 32, "height": 32},
  season_pin: {"url": "assets/wave8/time/season_pin.png", "width": 32, "height": 48},
  season_strip: {"url": "assets/wave8/time/season_strip.png", "width": 1024, "height": 48},
} as const;